+++ src/main/kotlin/team/incube/flooding/domain/neis/client/NeisTimetableClient.kt
@@ -45,4 +45,3 @@ class NeisTimetableClient(
             )
         }
 }
-
