+++ src/main/kotlin/team/incube/flooding/domain/neis/client/dto/GetMealsRequest.kt
@@ -1,0 +1,7 @@
+package team.incube.flooding.domain.neis.client.dto
+
+data class GetMealsRequest(
+    val officeCode: String,
+    val schoolCode: String,
+    val date: String,
+)
