--- src/main/kotlin/team/incube/flooding/domain/neis/client/dto/GetNeisMealsRequest.kt
@@ -1,7 +1,0 @@
-package team.incube.flooding.domain.neis.client.dto
-
-data class GetMealsRequest(
-    val officeCode: String,
-    val schoolCode: String,
-    val date: String,
-)
