--- src/main/kotlin/team/incube/flooding/domain/neis/client/dto/GetNeisTimetableRequest.kt
@@ -1,9 +1,0 @@
-package team.incube.flooding.domain.neis.client.dto
-
-data class GetTimetablesRequest(
-    val officeCode: String,
-    val schoolCode: String,
-    val grade: Int,
-    val classNumber: Int,
-    val date: String,
-)
