+++ src/main/kotlin/team/incube/flooding/domain/neis/presentation/data/response/GetNeisMealsResponse.kt
@@ -10,4 +10,3 @@ data class GetNeisMealsResponse(
         val calories: String?,
     )
 }
-
