+++ src/main/kotlin/team/incube/flooding/domain/neis/presentation/data/response/GetNeisTimetablesResponse.kt
@@ -13,4 +13,3 @@ data class GetNeisTimetablesResponse(
         val classroom: String?,
     )
 }
-
