+++ src/main/kotlin/team/incube/flooding/domain/neis/service/GetNeisTimetablesService.kt
@@ -1,0 +1,8 @@
+package team.incube.flooding.domain.neis.service
+
+import team.incube.flooding.domain.neis.presentation.data.request.GetNeisTimetablesRequest
+import team.incube.flooding.domain.neis.presentation.data.response.GetNeisTimetablesResponse
+
+interface GetNeisTimetablesService {
+    fun execute(request: GetNeisTimetablesRequest): GetNeisTimetablesResponse
+}
