+++ src/main/kotlin/team/incube/flooding/global/config/JacksonConfig.kt
@@ -1,0 +1,11 @@
+package team.incube.flooding.global.config
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class JacksonConfig {
+    @Bean
+    fun objectMapper(): ObjectMapper = ObjectMapper()
+}
